Move TypedElement property editor to gaphor.UML
It's a generic component used for different model elements.
